+++ src/odc_core/__init__.py
@@ -26,7 +26,7 @@ from . import (
     validate_username,
 )
 
-__version__ = "0.8.9"
+__version__ = "0.8.10"
 
 __all__ = [
     "browser_helpers",